git-svn-id: https://plugins.svn.wordpress.org/sitepress-multilingual-cms/trunk@140468 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ modules/icl-translation/icl-translation-functions.php
@@ -557,7 +557,9 @@ function icl_translation_delete_post($post_id){
     $wpdb->query("DELETE FROM {$wpdb->prefix}icl_node WHERE nid=".$post_id);
     $rid = $wpdb->get_var("SELECT rid FORM {$wpdb->prefix}icl_content_status WHERE nid=".$post_id);
     $wpdb->query("DELETE FROM {$wpdb->prefix}icl_content_status WHERE nid=".$post_id);
-    $wpdb->query("DELETE FROM {$wpdb->prefix}icl_core_status WHERE rid=".$rid);
+    if($rid){
+        $wpdb->query("DELETE FROM {$wpdb->prefix}icl_core_status WHERE rid=".$rid);
+    }
 }
 
 function icl_add_post_translation($trid, $translation, $lang, $rid){
